Update api for get create delete sla

+++ QuickServiceWebAPI/Services/Implements/SlaService.cs
@@ -9,11 +9,16 @@ namespace QuickServiceWebAPI.Services.Implements
     public class SlaService : ISlaService
     {
         private readonly ISlaRepository _repository;
+        private readonly ISlametricService _slametricService;
+        private readonly ISlametricRepository _slametricRepository;
         private readonly IMapper _mapper;
-        public SlaService(ISlaRepository repository, IMapper mapper)
+        public SlaService(ISlaRepository repository, IMapper mapper
+            , ISlametricService slametricService, ISlametricRepository slametricRepository)
         {
             _repository = repository;
             _mapper = mapper;
+            _slametricService = slametricService;
+            _slametricRepository = slametricRepository;
         }
 
         public List<SlaDTO> GetSLAs()
@@ -28,10 +33,27 @@ namespace QuickServiceWebAPI.Services.Implements
             return _mapper.Map<SlaDTO>(sla);
         }
 
-        public async Task<SlaDTO> CreateSLA(CreateUpdateSlaDTO createUpdateSlaDTO)
+        public async Task<SlaDTO> CreateSLA(CreateSlaDTO createSlaDTO)
         {
-            var sla = _mapper.Map<Sla>(createUpdateSlaDTO);
+            var sla = _mapper.Map<Sla>(createSlaDTO);
             sla.Slaid = await GetNextId();
+            using(var slaMetrics = sla.Slametrics.GetEnumerator())
+            {
+                string currentId = await _slametricService.GetNextId();
+                if (slaMetrics.MoveNext())
+                {
+                    slaMetrics.Current.SlametricId = currentId;
+                    slaMetrics.Current.Slaid = sla.Slaid;
+                }
+                while (slaMetrics.MoveNext())
+                {
+                    var nextId = GenerateNextIdForSlaMetrics(currentId);
+                    var slaMetric = slaMetrics.Current;
+                    slaMetric.SlametricId = nextId;
+                    slaMetric.Slaid = sla.Slaid;
+                    currentId = nextId;
+                }
+            }
             var slaAdded = await _repository.AddSLA(sla);
             if(slaAdded == null)
             {
@@ -40,21 +62,33 @@ namespace QuickServiceWebAPI.Services.Implements
             return _mapper.Map<SlaDTO>(slaAdded);
         }
 
-        public async Task UpdateSLA(string slaId, CreateUpdateSlaDTO createUpdateSlaDTO)
+        public async Task UpdateSLA(UpdateSlaDTO updateSlaDTO)
         {
-            Sla sla = await _repository.GetSLAById(slaId);
+            Sla sla = await _repository.GetSLAById(updateSlaDTO.SlaId);
             if (sla == null)
             {
                 throw new AppException("Sla not found");
             }
-            sla = _mapper.Map(createUpdateSlaDTO, sla);
+            sla = _mapper.Map(updateSlaDTO, sla);
             await _repository.UpdateSLA(sla);
         }
 
         public async Task DeleteSLA(string slaId)
         {
             var sla = await _repository.GetSLAById(slaId);
+            if(sla == null)
+            {
+                throw new AppException($"Sla with id {slaId} not found");
+            }
+            await _slametricRepository.DeleteSlaMetricsOfSla(sla);
             await _repository.DeleteSLA(sla);
+        }
+
+        private string GenerateNextIdForSlaMetrics(string currentId)
+        {
+            int id = IDGenerator.ExtractNumberFromId(currentId) + 1;
+            string slaId = IDGenerator.GenerateSlametricId(id);
+            return slaId;
         }
         public async Task<string> GetNextId()
         {
